further refactoring

+++ main_poll.c
@@ -26,6 +26,16 @@ bool revent_can_read(int revent) {
     || revent == POLLRDBAND || revent == POLLPRI;
 }
 
+bool revents_are_ok(struct pollfd* fds, size_t n) {
+  for(size_t i=0; i<n; ++i) {
+    if(!revent_is_ok(fds[i].revents)) return false;
+  }
+
+  return true;
+}
+
+const int N = 2;
+
 /* TODO: refactor it using for loops */
 int main(int argc, char** argv) {
   if(argc < 3) {
@@ -36,50 +46,35 @@ int main(int argc, char** argv) {
 
   int memfd = create_memfile_from(0, MEMFILENAME);
   if(memfd == -1) {
-    goto close_memfd;
+    return -1;
   }
 
-  int fd[2] = {
-    shm_open(MEMFILENAME, O_RDONLY),
-    shm_open(MEMFILENAME, O_RDONLY)
-  };
+  struct ProcessChild child[N];
+  struct pollfd fds[N];
+  struct Buffer buffer[N];
 
-  if(fd[0] == -1 || fd[1] == -1) {
-    perror("shm_open");
-    goto close_fd;
-  }
+  int initialized = 0;
+  for(; initialized < N; ++initialized) {
+    int fd = shm_open(MEMFILENAME, O_RDONLY, 0);
+    if(fd == -1) goto loop_open_error;
 
-  struct ProcessChild child[2] = {
-    spawn_process_child(argv[1], 0, fd[0], -1),
-    spawn_process_child(argv[2], 0, fd[1], -1),
-  };
+    child[initialized] = spawn_process_child(argv[initialized], 0, fd, -1);
+    if(child[initialized].status != 0) goto loop_child_error;
 
-  for(int i=0; i<2; ++i) {
-    if(child[i].running) continue;
-    goto kill_children;
-  }
-  
-  struct pollfd fds[2] = {
-    [0] = {
-      .fd = child[0].stdout,
-      .events = POLLIN,
-      .revents = 0,
-    },
-    [1] = {
-      .fd = child[1].stdout,
-      .events = POLLIN,
-      .revents = 0,
-    },
-  };
+    buffer[initialized] = buffer_with_capacity(4096);
+    if(buffer[initialized].ptr == 0) goto loop_buffer_error;
 
-  struct Buffer buffer[2] = {
-    buffer_with_capacity(4096),
-    buffer_with_capacity(4096),
-  };
+    fds[initialized].fd = child[initialized].stdout;
+    fds[initialized].events = POLLIN;
+    continue;
 
-  if(buffer[0].ptr == 0 || buffer[1].ptr == 0) {
+loop_buffer_error:
+    kill(child[initialized].pid, 9);
+loop_child_error:
+    close(fd);
+loop_open_error:
     goto destructors;
-  }
+  };
 
   do {
     int x = poll(fds, 2, 3000);
@@ -90,23 +85,17 @@ int main(int argc, char** argv) {
     }
 
     /* Timeout, check if something happened to children */
-    if(x == 0) for(int i=0; i<2; ++i) {
-      int pid = waitpid(child[i].pid, &child[i].status, WNOHANG);
-
-      /* Nothing happened */
-      if(pid == 0) continue; 
-
-      /* This should NOT happen */
-      if(pid == -1) {
-        perror("waitpid (after poll timeout)");
-        goto destructors;
-      }
-
-      child[i].running = false;
+    if(x == 0 && are_children_ok(child, N)) {
+      puts("Something happened to one of the children");
       goto destructors;
     }
 
-    for(int i=0; i<2; ++i) {
+    for(int i=0; i<N; ++i) {
+      if(!child[i].running) continue;
+      if(!is_child_ok(&child[i])) {
+        puts("One of the children returned with status other than 0");
+        goto destructors;
+      }
       if(!revent_can_read(fds[i].revents)) continue;
 
       ssize_t bytes_read = read_fd_to_buffer(fds[i].fd, &buffer[i]);
@@ -115,32 +104,36 @@ int main(int argc, char** argv) {
       perror("main read_fd_to_buffer");
       goto destructors;
     }
-  } while(revent_is_ok(fds[0].revents) || revent_is_ok(fds[1].revents));
+  } while(revents_are_ok(fds, N));
 
   if(buffer[0].size != buffer[1].size) {
-    puts("buffers dont have equal size");
+    puts("outputs dont have equal size");
 
-    char _msg1[] = "First:\n'";
+    /*
+    char _msg1[] = "First:\n";
     write(1, _msg1, sizeof(_msg1));
     write(1, buffer[0].ptr, buffer[0].size);
-    char _msg2[] = "'\nSecond:\n'";
+    char _msg2[] = "\nSecond:\n";
     write(1, _msg2, sizeof(_msg2));
     write(1, buffer[1].ptr, buffer[1].size);
-    char _msg3[] = "'\n";
+    char _msg3[] = "\n";
     write(1, _msg3, sizeof(_msg3));
+    */
 
     goto destructors;
   }
 
   if(memcmp(buffer[0].ptr, buffer[1].ptr, buffer[0].size) != 0) {
-    puts("buffers arent equal");
+    puts("outputs arent equal");
 
-    char _msg1[] = "First:\n'";
+    /*
+    char _msg1[] = "First:\n";
     write(1, _msg1, sizeof(_msg1));
-    char _msg2[] = "'\nSecond:\n'";
+    char _msg2[] = "\nSecond:\n";
     write(1, _msg2, sizeof(_msg2));
-    char _msg3[] = "'\n";
+    char _msg3[] = "\n";
     write(1, _msg3, sizeof(_msg3));
+    */
 
     goto destructors;
   }
@@ -148,23 +141,13 @@ int main(int argc, char** argv) {
   puts("yaaay");
 
 destructors:
-dealloc_buffers:
-  for(int i=0; i<2; ++i) {
+  for(int i=0; i<initialized; ++i) {
     munmap(buffer[i].ptr, buffer[i].capacity);
+    close(child[i].stdin);
+    if(child[i].running) kill(child[i].pid, 9);
   }
-
-kill_children:
-  for(int i=0; i<2; ++i) {
-    if(!child[i].running) continue;
-    kill(child[i].pid, 9);
-    waitpid(child[i].pid, 0, 0);
-  }
-close_fd:
-  for(int i=0; i<2; ++i) {
-    close(fd[i]);
-  }
-close_memfd:
   close(memfd);
+unlink_memfd:
   shm_unlink(MEMFILENAME);
 
   return 0;
